Fix: Implement hole collection pipeline for tetrahedral mesh generation
Added new column in the constraint table to define hole.
Added a new visualization for hole in the segmentation visualization
Added a synchronization system (currently flawed not fixed) to select a hole and it is selected in all other surfaces.
The build plc from selection have the new hole system which is passed down to the triangulator so no need to pass it down in the system to tetgen if its a hole. Will need to improve this part the logic is flawed